Communication_Protocol/requirements.txt
pytest
pytest-asyncio
pyserial
